+++ src/main/java/io/github/astrapi69/swing/combobox/model/AbstractComboBoxModel.java
@@ -35,9 +35,9 @@ import javax.swing.ComboBoxModel;
 import lombok.Getter;
 import lombok.Setter;
 import io.github.astrapi69.check.Argument;
-import io.github.astrapi69.collections.array.ArrayExtensions;
-import io.github.astrapi69.collections.list.ListExtensions;
-import io.github.astrapi69.collections.list.ListFactory;
+import io.github.astrapi69.collection.array.ArrayExtensions;
+import io.github.astrapi69.collection.list.ListExtensions;
+import io.github.astrapi69.collection.list.ListFactory;
 
 /**
  * The abstract class {@link AbstractComboBoxModel} contains the data for a combo list and the