+++ src/main/java/io/github/astrapi69/swing/list/JListExtensions.java
@@ -35,8 +35,8 @@ import javax.swing.ListSelectionModel;
 import javax.swing.border.EmptyBorder;
 import javax.swing.plaf.metal.MetalIconFactory;
 
-import io.github.astrapi69.collections.list.ListExtensions;
-import io.github.astrapi69.collections.list.ListFactory;
+import io.github.astrapi69.collection.list.ListExtensions;
+import io.github.astrapi69.collection.list.ListFactory;
 import io.github.astrapi69.model.check.CheckableItem;
 import io.github.astrapi69.model.check.CheckableValue;
 import io.github.astrapi69.swing.tree.renderer.list.CheckBoxListRenderer;