+++ src/test/java/io/github/astrapi69/swing/component/JMTextAreaTest.java
@@ -29,7 +29,7 @@ import java.awt.Frame;
 import javax.swing.JButton;
 
 import net.miginfocom.swing.MigLayout;
-import io.github.astrapi69.collections.pairs.ValueBox;
+import io.github.astrapi69.collection.pair.ValueBox;
 import io.github.astrapi69.model.LambdaModel;
 import io.github.astrapi69.window.adapter.CloseWindow;