+++ src/test/java/io/github/astrapi69/swing/panel/field/LabeledComboboxPanel.java
@@ -35,7 +35,7 @@ import javax.swing.JTextField;
 import javax.swing.LayoutStyle;
 
 import net.miginfocom.swing.MigLayout;
-import io.github.astrapi69.collections.list.ListFactory;
+import io.github.astrapi69.collection.list.ListFactory;
 import io.github.astrapi69.model.BaseModel;
 import io.github.astrapi69.model.api.IModel;
 import io.github.astrapi69.swing.base.BasePanel;