+++ src/test/java/io/github/astrapi69/swing/test/TestAutoBinding.java
@@ -36,7 +36,7 @@ import org.jdesktop.beansbinding.BeanProperty;
 import org.jdesktop.beansbinding.Bindings;
 import org.junit.jupiter.api.Test;
 
-import io.github.astrapi69.collections.pairs.ValueBox;
+import io.github.astrapi69.collection.pair.ValueBox;
 import io.github.astrapi69.test.object.Employee;
 import io.github.astrapi69.test.object.Person;